refactor: extract CustomerInvoiceDetailsTabsOptionsEnum
