+++ src/core/constants/NavigationEnum.ts
@@ -1,0 +1,4 @@
+export enum CustomerInvoiceDetailsTabsOptionsEnum {
+  overview = 'overview',
+  creditNotes = 'credit-notes',
+}
